backend/requirements.txt
fastapi
uvicorn
python-multipart

pdfplumber
python-docx

scikit-learn
torch
numpy
pandas

regex
spacy

matplotlib